Update WorldScaleSceneView tutorial
