+++ Sources/ArcGISToolkit/Documentation.docc/Resources/WorldScaleSceneView/WorldScaleSceneViewStep7.swift
@@ -20,17 +20,14 @@ struct WorldScaleExampleView: View {
         return scene
     }()
     
-    /// The location datasource that is used to access the device location.
-    @State private var locationDataSource = SystemLocationDataSource()
-    
-    /// The graphics overlay which shows a graphic around your initial location.
-    @State private var graphicsOverlay = GraphicsOverlay()
+    /// The world-tracking provider used by this example.
+    @State private var provider = AppleWorldTracking(mode: .worldTracking)
     
     var body: some View {
-        WorldScaleSceneView(
-            clippingDistance: 400
-        ) { _ in
-            SceneView(scene: scene, graphicsOverlays: [graphicsOverlay])
+        WorldScaleSceneView(provider: provider) { context in
+            AppleWorldTrackingCameraFeedView(context: context)
+        } sceneView: { _ in
+            SceneView(scene: scene)
         }
         .task {
             let locationManager = CLLocationManager()
@@ -38,31 +35,10 @@ struct WorldScaleExampleView: View {
                 locationManager.requestWhenInUseAuthorization()
             }
             
-            try? await locationDataSource.start()
+            try? await provider.dataSource.start()
             
             // Retrieve initial location.
-            guard let initialLocation = await locationDataSource.locations.first(where: { _ in true }) else { return }
-            
-            // Put a circle graphic around the initial location.
-            let circle = GeometryEngine.geodeticBuffer(
-                around: initialLocation.position,
-                distance: 20,
-                distanceUnit: .meters,
-                maxDeviation: 1,
-                curveType: .geodesic
-            )
-            graphicsOverlay.addGraphic(
-                Graphic(
-                    geometry: circle,
-                    symbol: SimpleLineSymbol(
-                        color: .red,
-                        width: 3
-                    )
-                )
-            )
-            
-            // Stop the location datasource after the initial location is retrieved.
-            await locationDataSource.stop()
+            guard let initialLocation = await provider.dataSource.locations.first(where: { @Sendable _ in true }) else { return }
         }
     }
 }
