+++ Sources/ArcGISToolkit/Documentation.docc/Resources/WorldScaleSceneView/WorldScaleSceneViewStep8.swift
@@ -20,6 +20,9 @@ struct WorldScaleExampleView: View {
         return scene
     }()
     
+    /// The graphics overlay which shows a graphic around your initial location.
+    @State private var graphicsOverlay = GraphicsOverlay()
+    
     /// The world-tracking provider used by this example.
     @State private var provider = AppleWorldTracking(mode: .worldTracking)
     
@@ -34,6 +37,29 @@ struct WorldScaleExampleView: View {
             if locationManager.authorizationStatus == .notDetermined {
                 locationManager.requestWhenInUseAuthorization()
             }
+            
+            try? await provider.dataSource.start()
+            
+            // Retrieve initial location.
+            guard let initialLocation = await provider.dataSource.locations.first(where: { @Sendable _ in true }) else { return }
+            
+            // Put a circle graphic around the initial location.
+            let circle = GeometryEngine.geodeticBuffer(
+                around: initialLocation.position,
+                distance: 20,
+                distanceUnit: .meters,
+                maxDeviation: 1,
+                curveType: .geodesic
+            )
+            graphicsOverlay.addGraphic(
+                Graphic(
+                    geometry: circle,
+                    symbol: SimpleLineSymbol(
+                        color: .red,
+                        width: 3
+                    )
+                )
+            )
         }
     }
 }
